# services/api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway for the Monitor SaaS platform",
        "title": "Monitor SaaS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://monitor-saas.com/support",
            "email": "support@monitor-saas.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/alerts": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all alert configurations for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "List alert configurations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ListAlertConfigsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new alert configuration for a monitor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "Create alert config",
                "parameters": [
                    {
                        "description": "Alert config details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateAlertConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GetAlertConfigResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/alerts/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get details for a specific alert config by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "Get alert config details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Alert Config ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GetAlertConfigResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update an existing alert configuration by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "Update alert config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Alert Config ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Alert config update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateAlertConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GetAlertConfigResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete an alert configuration by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "Delete alert config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Alert Config ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "types.AlertCondition": {
            "type": "string",
            "enum": [
                "status_code",
                "latency",
                "availability",
                "ssl_expiry",
                "keyword",
                "pattern"
            ],
            "x-enum-varnames": [
                "AlertConditionStatusCode",
                "AlertConditionLatency",
                "AlertConditionAvailability",
                "AlertConditionSSLExpiry",
                "AlertConditionKeyword",
                "AlertConditionPattern"
            ]
        },
        "types.AlertConfig": {
            "type": "object",
            "properties": {
                "condition": {
                    "$ref": "#/definitions/types.AlertCondition"
                },
                "consecutive_count": {
                    "type": "integer"
                },
                "cooldown_minutes": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "monitor_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "severity": {
                    "$ref": "#/definitions/types.AlertSeverity"
                },
                "threshold": {
                    "$ref": "#/definitions/types.AlertThreshold"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "types.AlertSeverity": {
            "type": "string",
            "enum": [
                "info",
                "warning",
                "critical"
            ],
            "x-enum-varnames": [
                "AlertSeverityInfo",
                "AlertSeverityWarning",
                "AlertSeverityCritical"
            ]
        },
        "types.AlertThreshold": {
            "type": "object",
            "properties": {
                "days_in_advance": {
                    "description": "For SSL expiry alerts",
                    "type": "integer"
                },
                "exact_match": {
                    "type": "string"
                },
                "max": {
                    "type": "number"
                },
                "min": {
                    "type": "number"
                },
                "pattern": {
                    "type": "string"
                }
            }
        },
        "types.CreateAlertConfigRequest": {
            "type": "object",
            "required": [
                "condition",
                "consecutive_count",
                "cooldown_minutes",
                "monitor_id",
                "name",
                "severity",
                "threshold"
            ],
            "properties": {
                "condition": {
                    "$ref": "#/definitions/types.AlertCondition"
                },
                "consecutive_count": {
                    "type": "integer",
                    "minimum": 1
                },
                "cooldown_minutes": {
                    "type": "integer",
                    "minimum": 1
                },
                "monitor_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "severity": {
                    "$ref": "#/definitions/types.AlertSeverity"
                },
                "threshold": {
                    "$ref": "#/definitions/types.AlertThreshold"
                }
            }
        },
...
        "types.GetAlertConfigResponse": {
            "type": "object",
            "properties": {
                "alert_config": {
                    "$ref": "#/definitions/types.AlertConfig"
                }
            }
        },
...
        "types.ListAlertConfigsResponse": {
            "type": "object",
            "properties": {
                "alert_configs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.AlertConfig"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "types.UpdateAlertConfigRequest": {
            "type": "object",
            "properties": {
                "condition": {
                    "$ref": "#/definitions/types.AlertCondition"
                },
                "consecutive_count": {
                    "type": "integer",
                    "minimum": 1
                },
                "cooldown_minutes": {
                    "type": "integer",
                    "minimum": 1
                },
                "enabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "severity": {
                    "$ref": "#/definitions/types.AlertSeverity"
                },
                "threshold": {
                    "$ref": "#/definitions/types.AlertThreshold"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "JWT token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}